@@ -1,19 +1,16 @@
-package com.eina.encryption.chat;
+package com.eina.chat.encryptionapi;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
 public class CryptoControllerTest {
@@ -22,7 +19,7 @@ public class CryptoControllerTest {
     private MockMvc mockMvc;
 
     @Test
-    public void symmetricCryptoTest() throws Exception{
+    public void symmetricCryptoTest(){
         try {
             String data = "dataTestForSymmetricEncrypt";
             MvcResult symEncrypted = mockMvc.perform(post("/symmetricEncrypt").content(data)
@@ -34,6 +31,7 @@ public class CryptoControllerTest {
                     .contentType(MediaType.TEXT_PLAIN)).andExpect(status().isOk()).andReturn();
             String decrypted = symDecrypted.getResponse().getContentAsString();
             System.out.println("Symmetic decrypted data: " + decrypted);
+            assert !data.equals(encrypted);
             assert data.equals(decrypted);
         } catch (Exception e){
             e.printStackTrace();
@@ -41,19 +39,25 @@ public class CryptoControllerTest {
     }
 
     @Test
-    public void asymmetricCryptoTest() throws Exception{
+    public void asymmetricCryptoTest(){
         try {
             String data = "dataTestForAAASymmetricEncrypt";
-            MvcResult symEncrypted = mockMvc.perform(post("/asymmetricEncrypt").content(data)
+            MvcResult asymEncrypted = mockMvc.perform(post("/asymmetricEncrypt").content(data)
                     .contentType(MediaType.TEXT_PLAIN)).andExpect(status().isOk()).andReturn();
-            String encrypted = symEncrypted.getResponse().getContentAsString();
+            String encrypted = asymEncrypted.getResponse().getContentAsString();
+
+            MvcResult asymEncrypted2 = mockMvc.perform(post("/asymmetricEncrypt").content(data)
+                    .contentType(MediaType.TEXT_PLAIN)).andExpect(status().isOk()).andReturn();
+            String encrypted2 = asymEncrypted2.getResponse().getContentAsString();
+
             System.out.println("Data to encrypt: " + data);
+
             System.out.println("ASymmetic encrypted data: " + encrypted);
-            MvcResult symDecrypted = mockMvc.perform(post("/asymmetricDecrypt").content(encrypted)
-                    .contentType(MediaType.TEXT_PLAIN)).andExpect(status().isOk()).andReturn();
-            String decrypted = symDecrypted.getResponse().getContentAsString();
-            System.out.println("ASymmetic decrypted data: " + decrypted);
-            assert data.equals(decrypted);
+
+            System.out.println("ASymmetic encrypted data 2: " + encrypted2);
+
+            assert !data.equals(encrypted);
+            assert encrypted.equals(encrypted2);
         } catch (Exception e){
             e.printStackTrace();
         }
